requirements.txt
pandas
faker
PyYAML
openpyxl
typer[all]
